+++ examples/collection/assembly/index.ts
@@ -1,6 +1,6 @@
 import { collections } from "@hypermode/functions-as";
 import { models } from "@hypermode/functions-as";
-import { EmbeddingsModel } from "@hypermode/models-as/models/openai/embeddings";
+import { OpenAIEmbeddingsModel } from "@hypermode/models-as/models/openai/embeddings";
 
 // These names should match the ones defined in the hypermode.json manifest file.
 const modelName: string = "embeddings";
@@ -9,10 +9,10 @@ const searchMethods: string[] = ["searchMethod1", "searchMethod2"];
 
 // This function takes input text and returns the vector embedding for that text.
 export function embed(text: string): f64[] {
-  const model = models.getModel<EmbeddingsModel>(modelName);
+  const model = models.getModel<OpenAIEmbeddingsModel>(modelName);
   const input = model.createInput(text);
   const output = model.invoke(input);
-  return output.data[0].embedding;
+  return output.data[0].embedding.map<f64>((e) => e); // TODO, remove the map when we are allowing f32[] return type
 }
 
 export function addProduct(description: string): string {